requirements.txt
ccxt==4.3.89
pandas==2.2.2
pandas_ta==0.3.14b
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.3
